src/main/java/com/doityourself/workshop/features/project/material/exceptionHandler/SaveProjectMaterialGlobalExceptionHandler.java
package com.doityourself.workshop.features.project.material.exceptionHandler;

import com.doityourself.workshop.common.command.CommandDTO;
import com.doityourself.workshop.common.command.ICommandExceptionHandler;
import com.doityourself.workshop.features.project.material.exception.SaveProjectMaterialFailedException;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

import java.util.Collections;

/**
 * Save Project Material Global Exception Handler
 */
@Slf4j
@Component
public class SaveProjectMaterialGlobalExceptionHandler implements ICommandExceptionHandler<CommandDTO> {
  @Value("${project.save.material.validation.failed}")
  String projectSaveMaterialFailedErrorMessage;

  /**
   * Method to handle runtime exception
   *
   * @param dto {@link CommandDTO}
   * @param runtimeException {@link RuntimeException}
   */
  @Override
  public void handleRuntimeException(CommandDTO dto, RuntimeException runtimeException) {
    handle();
  }

  /**
   * Method to handle exception
   *
   * @param dto {@link CommandDTO}
   * @param exception {@link Exception}
   */
  @Override
  public void handleException(CommandDTO dto, Exception exception) {
    handle();
  }

  /**
   * Common exception handler
   */
  private void handle() {
    SaveProjectMaterialFailedException exception = new SaveProjectMaterialFailedException();
    exception.setMessages(Collections.singletonList(projectSaveMaterialFailedErrorMessage));
    throw exception;
  }
}
